+++ Core/Samples/inc/MultiLayer.h
@@ -89,7 +89,7 @@ public:
     //double getCrossCorrFun(const kvector_t &k, int j, int k) const;
 
     //! fourier transform of the correlation function of roughnesses between the interfaces
-    double getCrossCorrSpectralFun(const kvector_t &kvec, int j, int k) const;
+    double getCrossCorrSpectralFun(const kvector_t &kvec, size_t j, size_t k) const;
 
     //! change thickness of layer
     void setLayerThickness(size_t i_layer, double thickness);
